Fixups following module renaming
