+++ cardinal_pythonlib/email_utils/mailboxpurge.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# cardinal_pythonlib/email/mailboxpurge.py
+# cardinal_pythonlib/email_utils/mailboxpurge.py
 
 """
 Remove all binary attachments from email messages
